added pages folder and changed file structure in client
