+++ client/src/pages/Detail.js
@@ -14,7 +14,7 @@ import {
   MDBCardBody, 
 } from 'mdbreact';
 
-import Navbar from "./Navbar";
+import Navbar from "../components/Navbar";
 
 const Detail = (props) => {
   const [recipe, setRecipe] = useState({});
